chore: rename styled date picker
